Some more fixes

+++ xdg/mutt-mailto.py
@@ -11,7 +11,7 @@ import os
 # or
 # mailto://user@domain?subject=XYZ&cc=...&body=...
 
-print(sys.argv)
+#print(sys.argv)
 url   = sys.argv[1]
 o     = urllib.parse.urlparse(url)
 query = urllib.parse.parse_qs(o.query)
@@ -24,27 +24,31 @@ address = o.netloc + o.path # one of them is empty
 # >>> urlparse('mailto://bla@blub.com')
 # ParseResult(scheme='mailto', netloc='bla@blub.com', path='', params='', query='', fragment='')
 
-draft = tempfile.NamedTemporaryFile(mode='w',delete=False)
+draft = None
 if 'body' in query:
+    draft = tempfile.NamedTemporaryFile(mode='w',delete=False)
     draft.write(query['body'][0].replace("\r","") + '\n')
-draft.close()
+    draft.close()
 
 
 proc = [ ]
 #proc += [ "urxvt", "-e" ]
-proc += [ "mutt" ]
+proc += [ "mutt"]
 if 'subject' in query:
     proc += [ '-s', query['subject'][0] ]
 if 'cc' in query:
     for cc in query['cc']:
         proc += [ '-c', cc ]
-proc += [ '-H', draft.name ]
+if draft is not None:
+    proc += [ '-H', draft.name ]
 proc += [ '--', address ]
 
+print(' '.join(proc), file=sys.stderr)
 mutt = subprocess.Popen(proc) #, stdin=subprocess.PIPE)
 mutt.wait()
 #print(draft.name)
-os.unlink(draft.name)
+if draft is not None:
+    os.unlink(draft.name)
 #print (proc)
 
 #if 'body' in query:
